pyproject.toml
[project]
name = "yoi-yurei"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pygame-ce>=2.5.6",
    "pyinstaller>=6.16.0",
]
